adding repo for stockinit and product

+++ microservices-api/spring-reactive-mongoapi/src/main/java/com/inti/formation/shop/api/repository/model/Stockinit.java
@@ -27,7 +27,7 @@ public class Stockinit implements Serializable {
     @Indexed(unique = false)
     private String magasin;
     private long quantite;
-    private int idproduct;
+    private long idproduct;
     private boolean active;
 
     @Transient
